Upload R version of intro tutorial
The main tutorial file is a `.qmd` file is a Quarto document that can be run like a normal RMarkdown document. This can be used interactively like a notebook, or rendered into a document (e.g., the .html file also found here). The fitModel.R and simulateData.R files contain functions that are used in the main tutorial file.